+++ Binar/Binar/MF/Animal/Cell/MFAnimalTableCell.xib
@@ -1,0 +1,62 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="20037" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES">
+    <device id="retina6_1" orientation="portrait" appearance="light"/>
+    <dependencies>
+        <deployment identifier="iOS"/>
+        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="20020"/>
+        <capability name="Safe area layout guides" minToolsVersion="9.0"/>
+        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
+    </dependencies>
+    <objects>
+        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner"/>
+        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
+        <tableViewCell contentMode="scaleToFill" selectionStyle="default" indentationWidth="10" reuseIdentifier="MFAnimalTableCell" rowHeight="132" id="KGk-i7-Jjw" customClass="MFAnimalTableCell" customModule="Binar" customModuleProvider="target">
+            <rect key="frame" x="0.0" y="0.0" width="320" height="132"/>
+            <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
+            <tableViewCellContentView key="contentView" opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center" tableViewCell="KGk-i7-Jjw" id="H2p-sc-9uM">
+                <rect key="frame" x="0.0" y="0.0" width="320" height="132"/>
+                <autoresizingMask key="autoresizingMask"/>
+                <subviews>
+                    <imageView clipsSubviews="YES" userInteractionEnabled="NO" contentMode="scaleAspectFit" horizontalHuggingPriority="251" verticalHuggingPriority="251" translatesAutoresizingMaskIntoConstraints="NO" id="0ku-kf-sxL">
+                        <rect key="frame" x="16" y="16" width="100" height="100"/>
+                        <constraints>
+                            <constraint firstAttribute="width" constant="100" id="O2i-aP-NOt"/>
+                            <constraint firstAttribute="height" constant="100" id="mcr-YH-5dA"/>
+                        </constraints>
+                    </imageView>
+                    <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Label" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="IKa-BR-Pov">
+                        <rect key="frame" x="132" y="16" width="172" height="21"/>
+                        <fontDescription key="fontDescription" type="system" weight="medium" pointSize="17"/>
+                        <nil key="textColor"/>
+                        <nil key="highlightedColor"/>
+                    </label>
+                    <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Label" textAlignment="natural" lineBreakMode="tailTruncation" numberOfLines="4" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="gUE-CP-tkc">
+                        <rect key="frame" x="132" y="45" width="172" height="18"/>
+                        <fontDescription key="fontDescription" type="system" pointSize="15"/>
+                        <nil key="textColor"/>
+                        <nil key="highlightedColor"/>
+                    </label>
+                </subviews>
+                <constraints>
+                    <constraint firstAttribute="trailing" secondItem="gUE-CP-tkc" secondAttribute="trailing" constant="16" id="5Zc-DA-ZfK"/>
+                    <constraint firstItem="IKa-BR-Pov" firstAttribute="leading" secondItem="0ku-kf-sxL" secondAttribute="trailing" constant="16" id="DS9-t7-Qo5"/>
+                    <constraint firstAttribute="bottom" secondItem="0ku-kf-sxL" secondAttribute="bottom" constant="16" id="LMh-zT-Mqi"/>
+                    <constraint firstAttribute="bottom" relation="greaterThanOrEqual" secondItem="gUE-CP-tkc" secondAttribute="bottom" constant="16" id="Sbo-0x-krl"/>
+                    <constraint firstItem="0ku-kf-sxL" firstAttribute="leading" secondItem="H2p-sc-9uM" secondAttribute="leading" constant="16" id="Sdj-V1-ZFd"/>
+                    <constraint firstItem="IKa-BR-Pov" firstAttribute="top" secondItem="H2p-sc-9uM" secondAttribute="top" constant="16" id="XCf-tq-NRJ"/>
+                    <constraint firstItem="0ku-kf-sxL" firstAttribute="top" secondItem="H2p-sc-9uM" secondAttribute="top" constant="16" id="XjO-c9-87x"/>
+                    <constraint firstAttribute="trailing" secondItem="IKa-BR-Pov" secondAttribute="trailing" constant="16" id="dp7-bO-qCh"/>
+                    <constraint firstItem="gUE-CP-tkc" firstAttribute="leading" secondItem="0ku-kf-sxL" secondAttribute="trailing" constant="16" id="jQO-Pr-py2"/>
+                    <constraint firstItem="gUE-CP-tkc" firstAttribute="top" secondItem="IKa-BR-Pov" secondAttribute="bottom" constant="8" id="o1a-Y8-LCg"/>
+                </constraints>
+            </tableViewCellContentView>
+            <viewLayoutGuide key="safeArea" id="njF-e1-oar"/>
+            <connections>
+                <outlet property="imageViewCell" destination="0ku-kf-sxL" id="aqo-iI-yzg"/>
+                <outlet property="subtitleLabel" destination="gUE-CP-tkc" id="LDR-ly-WSh"/>
+                <outlet property="titleLabel" destination="IKa-BR-Pov" id="5CS-TR-Pge"/>
+            </connections>
+            <point key="canvasLocation" x="131.8840579710145" y="107.14285714285714"/>
+        </tableViewCell>
+    </objects>
+</document>
